refactor(wordle): rename folders
