+++ src/app/(pages)/games/wordleverse/_components/index.js
